custom exception handler

+++ src/main/java/com/example/springquiz/service/impl/AnswerService.java
@@ -1,10 +1,12 @@
 package com.example.springquiz.service.impl;
 
 import com.example.springquiz.builder.AnswerBuilder;
-import com.example.springquiz.exception.CustomizedNotFoundException;
+import com.example.springquiz.enumeration.ErrorCode;
+import com.example.springquiz.exception.CustomizedRuntimeException;
 import com.example.springquiz.model.domain.Answer;
 import com.example.springquiz.model.dto.AnswerDTO;
 import com.example.springquiz.repository.IAnswerRepository;
+import com.example.springquiz.repository.IQuestionRepository;
 import com.example.springquiz.service.IAnswerService;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -19,6 +21,7 @@ import java.util.stream.Stream;
 public class AnswerService implements IAnswerService {
     private final IAnswerRepository answerRepository;
     private final AnswerBuilder answerBuilder;
+    private final IQuestionRepository questionRepository;
 
     @Override
     public int createNewAnswer(AnswerDTO dto) {
@@ -38,11 +41,14 @@ public class AnswerService implements IAnswerService {
     public Optional<AnswerDTO> getAnswerById(int id) {
         return answerRepository.findById(id)
                 .map(answerBuilder::build)
-                .orElseThrow(() -> new CustomizedNotFoundException(String.format("No such Answer for id '%s'", id)));
+                .orElseThrow(() -> new CustomizedRuntimeException(ErrorCode.ANSWER_NOT_FOUND));
     }
 
     @Override
     public List<Optional<AnswerDTO>> getAnswersByQuestionId(int questionId) {
+        if(questionRepository.findById(questionId).isEmpty()) {
+            throw new CustomizedRuntimeException(ErrorCode.QUESTION_NOT_FOUND);
+        }
         return answerRepository.findByQuestion_QuestionId(questionId)
                 .stream()
                 .map(answerBuilder::build)
@@ -55,11 +61,14 @@ public class AnswerService implements IAnswerService {
                 .map(model -> answerBuilder.build(dto, model))
                 .map(answerRepository::save)
                 .map(answerBuilder::build)
-                .orElseThrow(() -> new CustomizedNotFoundException(String.format("No such Answer for id '%s'", id)));
+                .orElseThrow(() -> new CustomizedRuntimeException(ErrorCode.ANSWER_NOT_FOUND));
     }
 
     @Override
     public void deleteAnswerById(int id) {
+        if(answerRepository.findById(id).isEmpty()) {
+            throw new CustomizedRuntimeException(ErrorCode.ANSWER_NOT_FOUND);
+        }
         answerRepository.deleteById(id);
     }
